Fixes issues due to modules move.
