+++ papertty/drivers/drivers_full.py
@@ -15,7 +15,8 @@
 #     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from abc import abstractmethod
-from papertty.drivers import WaveshareEPD
+
+from papertty.drivers.drivers_base import WaveshareEPD
 
 
 class WaveshareFull(WaveshareEPD):
